+++ src/hassette/models/states/alarm_control_panel.py
@@ -1,4 +1,4 @@
-from typing import Any, Literal
+from typing import Literal
 
 from pydantic import Field
 
@@ -7,10 +7,8 @@ from .base import AttributesBase, StringBaseState
 
 class AlarmControlPanelAttributes(AttributesBase):
     code_format: str | None = Field(default=None)
-    changed_by: Any | None = Field(default=None)
+    changed_by: str | None = Field(default=None)
     code_arm_required: bool | None = Field(default=None)
-    previous_state: Any | None = Field(default=None)
-    next_state: Any | None = Field(default=None)
 
 
 class AlarmControlPanelState(StringBaseState):